requirements.txt
anthropic==0.30.0
ipython==8.26.0
httpx==0.27.0
python-dotenv==1.0.1